init: refactored view keys
Wanted to use the same keys for changing both the buffer and view.
In doing so, made a pair of connect.events to accomplish the task.
It works with only `BUFFER_NEW`, but if missing the `VIEW_NEW`, then
a view split would need a new buffer to initialize the keymap. Instead,
using the `VIEW_NEW` allows for auto initalization and automatically
disables buffer change keymap.